+++ test/e2e/tests/bridge/swap-positive-cases.spec.ts
@@ -1,10 +1,11 @@
 import { strict as assert } from 'assert';
 import { Suite } from 'mocha';
 import { getEventPayloads, withFixtures } from '../../helpers';
+import { bridgeTransaction } from '../../page-objects/flows/bridge.flow';
 import { login } from '../../page-objects/flows/login.flow';
 import HomePage from '../../page-objects/pages/home/homepage';
 import { BRIDGE_FEATURE_FLAGS_WITH_SSE_ENABLED } from './constants';
-import { bridgeTransaction, getBridgeFixtures } from './bridge-test-utils';
+import { getBridgeFixtures } from './bridge-test-utils';
 
 describe('Swap tests', function (this: Suite) {
   this.timeout(160000); // This test is very long, so we need an unusually high timeout
