+++ test/e2e/tests/swaps/swap-erc20-with-loaded-state.spec.ts
@@ -1,9 +1,7 @@
 import { withFixtures } from '../../helpers';
 import { login } from '../../page-objects/flows/login.flow';
-import {
-  bridgeTransaction,
-  getBridgeFixtures,
-} from '../bridge/bridge-test-utils';
+import { bridgeTransaction } from '../../page-objects/flows/bridge.flow';
+import { getBridgeFixtures } from '../bridge/bridge-test-utils';
 import { DEFAULT_BRIDGE_FEATURE_FLAGS } from '../bridge/constants';
 
 describe('Swap', function () {
